requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.0.5
pdfkit==1.0.0
reportlab==4.0.4
python-dotenv==1.0.1
Flask-JWT-Extended==4.6.0
Flask-Cors==5.0.0
psycopg[binary]==3.2.9
openpyxl==3.1.5
Flask-Talisman==1.1.0
Flask-Limiter==3.9.2

python-docx==1.1.0
Flask-Compress==1.15
alembic==1.13.2